+++ Ecifircas/Ecifircas/perft.h
@@ -4,6 +4,6 @@
 
 namespace Ecifircas
 {
-	extern long Nodes;
+	extern long long Nodes;
 	void perft_driver(int depth);
 }
